requirements.txt
torch>=2.2
torchvision>=0.17
numpy
pandas
scikit-learn
faiss-cpu
tqdm
Pillow
matplotlib
PyYAML
flask
